requirements.txt
discordwebhook==1.0.3
matplotlib==2.2.2
pandas==1.1.5
pytesseract==0.3.7
python-telegram-bot==13.7
PyYAML==5.4.1
requests==2.25.1
schedule==1.1.0
tesseract==0.1.3
Pillow >= 1.0